+++ src/parser/expressions/sub.rs
@@ -1,0 +1,20 @@
+use crate::parser::parse::BoxedExpression;
+use crate::parser::{Error, Expression, Result, Value};
+
+#[derive(Debug)]
+pub(in crate::parser) struct Sub {
+    pub left: BoxedExpression,
+    pub right: BoxedExpression,
+}
+
+impl Expression for Sub {
+    fn calculate(&self, json: &[u8]) -> Result<Value> {
+        let left = self.left.calculate(json)?;
+        let right = self.right.calculate(json)?;
+
+        match (left, right) {
+            (Value::Number(n1), Value::Number(n2)) => Ok(Value::Number(n1 - n2)),
+            (l, r) => Err(Error::UnsupportedTypeComparison(format!("{l} - {r}",))),
+        }
+    }
+}
